Re-Mix and Bug Fixing
Lowered the volume of most sounds, and the rate at which the wolf plays

I also tied the body sound effect to the hole barrier instead of the hole's mesh, so it doesn't play multiple times. The cleaning sound also works as intended now, it's bound to the progress bar.